+++ packages/constants/src/utils.ts
@@ -1,0 +1,26 @@
+import { ChainId, Primitive, useChainId } from '@masknet/web3-shared'
+import { useMemo } from 'react'
+
+export interface Constants {
+    [K: string]: { [K in keyof typeof ChainId]: Primitive | Primitive[] }
+}
+
+export function transform<T extends Constants>(constants: T) {
+    return (chainId = ChainId.Mainnet) => {
+        const table = {} as { [key in keyof T]: T[key]['Mainnet'] }
+        const chainName = ChainId[chainId] as keyof typeof ChainId
+        for (const name in constants) {
+            const key = name as keyof T
+            table[key] = constants[key][chainName]
+        }
+        return Object.freeze(table)
+    }
+}
+
+export function hookTransform<T>(getConstants: (chainId: ChainId) => T) {
+    return function useConstants(chainId = ChainId.Mainnet) {
+        const current = useChainId()
+        const finalChain = chainId ?? current
+        return useMemo(() => getConstants(finalChain), [finalChain])
+    }
+}
